src/main/java/com/masiv/movies/controller/TicketController.java
package com.masiv.movies.controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.masiv.movies.models.Ticket;
import com.masiv.movies.service.TicketService;

@RestController
@RequestMapping("/ticket")
public class TicketController {
	@Autowired
	private TicketService ticketService;
	@PostMapping("/purchase")
    public ResponseEntity<String> purchaseTicket(@RequestBody Ticket ticket) {
        try {
            Ticket created = ticketService.buyTicket(ticket);
            
            return new ResponseEntity<>("Purchase successful.\n Ticket: " + created.getPurchaseId().toString(), HttpStatus.CREATED);
        } catch (IllegalArgumentException e) {
            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
        } catch (Exception e) {
            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
        }
    }
}